backend/requirements.txt
Flask==2.3.2
google-cloud-firestore==2.11.0
google-cloud-vision==3.5.0
google-generativeai==0.6.0
gunicorn==21.2.0